Suggestion

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -3458,7 +3458,7 @@ void ObjectMgr::LoadPlayerInfo()
                     continue;
                 }
 
-                int32 amount = fields[3].GetInt8();
+                int32 amount = fields[3].GetInt32();
 
                 if (!amount)
                 {
